# Replace custom CRC code with Crc32.NET lib for 30% speedup.