--- a/src/IronSnappy.Test/GoldenTest.cs
+++ b/src/IronSnappy.Test/GoldenTest.cs
@@ -55,5 +55,41 @@ namespace IronSnappy.Test
          Assert.Equal(bytes.Length, uncompressed.Length);
          Assert.Equal(bytes, uncompressed);
       }
+
+      [Fact]
+      public void WriteGoldenStream()
+      {
+         // Created with `cat Mark.Twain-Tom.Sawyer.txt | snappy -c > Mark.Twain-Tom.Sawyer.txt.snappy`
+         byte[] want = File.ReadAllBytes("TestData/Mark.Twain-Tom.Sawyer.txt.snappy");
+
+         using(FileStream infile = File.OpenRead("TestData/Mark.Twain-Tom.Sawyer.txt"))
+         using(MemoryStream memoryStream = new MemoryStream())
+         using(Stream snappyWriter = Snappy.OpenWriter(memoryStream))
+         {
+            infile.CopyTo(snappyWriter);
+            snappyWriter.Flush();
+            byte[] got = memoryStream.GetBuffer();
+
+            Assert.Equal(want.Length, got.Length);
+            Assert.Equal(want, got);
+         }
+      }
+
+      [Fact]
+      public void ReadGoldenStream()
+      {
+         byte[] want = File.ReadAllBytes("TestData/Mark.Twain-Tom.Sawyer.txt");
+
+         using(FileStream infile = File.OpenRead("TestData/Mark.Twain-Tom.Sawyer.txt.snappy"))
+         using(Stream snappyReader = Snappy.OpenReader(infile))
+         using(MemoryStream memoryStream = new MemoryStream())
+         {
+            snappyReader.CopyTo(memoryStream);
+            byte[] got = memoryStream.GetBuffer();
+
+            Assert.Equal(want.Length, got.Length);
+            Assert.Equal(want, got);
+         }
+      }
    }
 }