--- a/src/IronSnappy/IronSnappy.csproj
+++ b/src/IronSnappy/IronSnappy.csproj
@@ -26,6 +26,7 @@
    </PropertyGroup>
 
    <ItemGroup>
+      <PackageReference Include="Crc32.NET" Version="1.2.0" />
       <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
    </ItemGroup>