--- a/src/IronSnappy/SnappyReader.cs
+++ b/src/IronSnappy/SnappyReader.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using System.Buffers;
-using System.Collections.Generic;
 using System.IO;
-using System.Text;
 
 namespace IronSnappy
 {