… src/IronSnappy.Test/TestData/Mark.Twain-Tom.Sawyer.txt.snappy