@aloneguid Would you consider taking a dependency https://github.com/force-net/Crc32.NET in IronSnappy?
- This lib has ~10x faster CRC calculation.
- This equated to ~30% faster compression/decompression in IronSnappy.
- Gives identical results to IronSnappy implementation

Though unfortunately:
- Hasn't been updated in 3 years (but the code is quite small and self contained)

---

This PR is dependent on https://github.com/aloneguid/IronSnappy/pull/9, so should only be submitted after they are pulled.
The first to changesets in this PR are identical to #9 

I also did some benchmarking (results in https://github.com/rodo-r2r/IronSnappy/blob/rodo/FastCrc-benchmark/src/IronSnappy.Benchmark/PerfTest.cs#L10) comparing IronSnappy with the original vs Crc32.NET CRC library (and also against gzip)

Decoding a ~500mb binary file shows a 30% speedup (1.59s vs 2.34s). Decompression results are similar.
```
|                Method |        Mean |     Error |   StdDev | Ratio | RatioSD |       Gen 0 | Gen 1 | Gen 2 |     Allocated |
|       DecodeIronBench |  2,347.7 ms |   9.89 ms |  9.25 ms |  2.10 |    0.01 |           - |     - |     - |  2687385600 B |  # -33%
|       DecodeFastBench |  1,592.0 ms |   7.33 ms |  6.49 ms |  1.43 |    0.00 |  64000.0000 |     - |     - |  3227353520 B |  # -34%
|       DecodeGzipBench |  1,809.2 ms |  14.56 ms | 13.62 ms |  1.62 |    0.01 |           - |     - |     - |  2687123632 B |
```