@rodneyodonnell 2020-11-03
I'm abandoning this PR.

This approach seems to work ok, but requires an extra dependency and needs a bit more testing.
For my use case I need something a bit faster, so am switching to https://github.com/brantburnett/Snappier